Added max length handling of msg in case of PR status change
We have to keep the message that exceeds 140 characters in case of comment creation of issue and PR
In case that cibot create deployment statuses for a given PR, keep Maximum length of 140 characters.
But, we have to trim the message that exceeds 140 characters in case of PR status change.
* Default: ""
* https://developer.github.com/v3/repos/deployments/

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>