[Enhancement] Added exception folder into file size check routine
This PR is to append a folder name (e.g, temporal-bin) to exclude check routine of
file size limit. It means that +5 MB file (by default) can be uploaded temporarily
in case of an inevitable situation.

Basically you can not upload a big file that exceeds specified file size.
If you have to temporarily upload binary files unavoidably, a contributor
has to share the issue to all members after uploading the files in **specified folder**.

**Changes proposed in this PR:**
1. Added exception folder routine for file size limit
2. Updated report format
3. Updated configuration file.
4. Updated doxygen rule to handle $SKIP_CI_PATHS

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>