Fixed license issue that is resulted from scancode
*Resolves: issue #138

This commit is to fix cleanly possible license issues of open-source compliance
that is generated by Scancode.

ScanCode Toolkit is a set of code scanning tools to detect the origin and license
of code and dependencies. It uses a plug-in architecture to run a series of
scan-related tools in one process flow.

**Changes proposed in this PR:**
1. Fixed possible open-source license issues

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>