Yocto: Exception handling in case that Yocto eSDK is not installed
Note that administrator of a server has to specify the location of eSDK at first".
And then, user has to change the BUILD_MODE from 99 to 0.

1) Yocto: BUILD_MODE = 99 (skip, by default)
Skipping 'devtool' procedure temporarily because BUILD_MODE is 99.
'777' will be used for a fine-graind classification when the values of 'build_result' are increased.
$build_result=777

2) eSDK:  $YOCTO_ESDK_ROOT == ""  (Administrator has to install eSDK/Yocto)
Skipping 'devtool' procedure temporarily because YOCTO_ESDK_ROOT is empty."
In this case, ask administrator of a server to install Yocto eSDK.
'888' will be used for a fine-graind classification when the values of 'build_result' are increased.
$build_result=888

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>