config: Added configuration file for server administrator
Fixes issue #210.

This commit is to append a configuration file for a server administrator. The server
administrator has to install required packages and development environment for CI
server. As we all know, an initial setting is not done by administrator.
So, let's support a configirstion file for the server administrator.

* file: ./config/config-sercer-administrator.sh

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>