Feature: Added a module to check automatically misspelled statement
Fixes issue #168.

This commit is to append a facility in order to check a misspelled characters automatically
in the document files such as .md and .txt. Actually, the reviewing the typo(s) in a commit file
is not productive review activity.

** Proposal: **
1. Module Name:   pr-format-misspelling.sh
2. Plugins Category   : ./plugins-good/ group
3. Goal of this module:
Check a misspelled statement in a text document file with GNU Aspell

GNU Aspell is a Free and Open Source spell checker designed to eventually replace Ispell.
It can either be used as a library or as an independent spell checker. Its main feature
is that it does a superior job of suggesting possible replacements for a misspelled word
than just about any other spell checker out there for the English language. Unlike Ispell,
Aspell can also easily check documents in UTF-8 without having to use a special dictionary.

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>