doc: update apt repository of Ubuntu distro
This commit is to append how to do a self assessment for Ubuntu packaging work.
Note that contributors have to execute a self assessment before submitting your PR.

We describe how to verify if your commit(s) is valid. To complete a source code verification,
we strongly recommend that contributors execute a clean build with a latest version of
a software platform. A partial validation inspection of software cause potential defects
when the software is integrated into the software platform.

* For APT repository of Tensorflow and NNstreamer
deb [trusted=yes] http://ppa.launchpad.net/nnstreamer/ppa/ubuntu xenial main

**Changes proposed in this PR:**
1. Updated how to setup TAOS-CI server
2. Updated self assessment manaul
3. Updated guide documents

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>