NNStreamer module: rename a module of apptest
This commit is to rename the existing module name for nnstreamer apptest.
* before:pr-{module-name}.sh
* after :pr-{type}-{module-name}.sh

* Guide:

First of all, **Note** that you have to run the below statement in order to check the grammar error of a CI module that your write.
```bash
$ bash ./pr-{type}-{module-name}.sh
```

When you submit a pull request to merge your CI module to TAOS-CI repository, Please note that reviewers check 4 requirements
as follows before merging your pull request.
* Maintenance: The module has to be normally executed as a CI component after enabling the module via a configuration file.
* Readability: Most of the developer should be able to read source code of the CI module without difficulty.
* Execution time: The time required to execute the module should not be long.
* Compatibility: The module should be able to run on most Linux distributions as well as Ubuntu.

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>