PR gateway: Run the format (prog1) and audit (prog2) group in parallel.
Fixed issue #483.

This commit is to improve the exsiting PR gateway that distributes tasks..
When a webhook handler receive the event messages from the PRs,
The webhook handler call the PR gateway. Then, the PR gateway runs
the format (prog1) and audit (prog2) group in parallel.
Prog1 and Prog2 woudl be started in the background.

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>