NNstreamer/Apptest: Improve debug message to monitor wget status
The apptest module for nnstreamer, downloads a file from the specified
URL at first. Then, it archives the files /tmp folder. It means that
the download files are continually archived in /tmp folder until the
server will be rebooted.

This commit is to display exactly execution status of 'wget' command
between a download mode (first) and caching mode (from second).

**Changelog**
1. Replaced wget.log with wget_status.oog
2. Update debug messages
3. Separated the two statuses such as a download and caching mode.

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>