Changed a extension format of log files (.log to .txt)
This commit is to replace .log with .txt in order to improve compatibility
and readability when press the log files among the browsers. The www-data
of Apache web-server executes the scripts based on the configuration file
of Apache httpd daemon. In case of text files, let's use .txt format instead
of the .log file extension.

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>