doc: Added how to write a changelog
This commit is to describe how to write a changelog when a commit is updated.

**Keep a changelog**
A changelog is a file which contains a chronologically ordered list of notable
changes for each version of a commit. I has to be written to make it easier for
users and contributors in order to see precisely what notable changes have been
made between each version of the commit.
When a commit changes, reviewers want to know why and how.

**Changelog**
- Added how to write a changelog
- Removed the reference section (Deprecated)

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>