coverity module: improved performance of the coverity scan
This commit is to improve the execution time of the coverity scan
when  the current time is not appropriate time to build and commit the
nnstreamer source codes to the Coverity Scan Cloud (e.g., https://scan.coverity.com)

Currently, the default execution frequency is 12 hours for nnstreamer repository.
As a basic policy. We do not use cron daemon to run the build & commit task.
The execution time is dependent on the PR frequency that is submitted by contributor.

**Changelog**
   * Fixed typos
   * Enabled on/off step of the build/commit tasks
   * Added debug message for maintenance

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>