post-build-tizen/bugfix: Remove 'tee' connected to 'gbs build' via '|'
When the BUILD_MODE is set to 1, a command pipeline is used for running
'gbs build' and capturing the standard output of it using 'tee', which
could return '0' even if the former 'gbs build' command fails. To fix
this bug, this patch replaces the pipe, '|', between 'gbs build' and
'tee' with the redirections and removes 'tee'.

Signed-off-by: Wook Song <wook16.song@samsung.com>