[Doxygen] Add option to skip definition tag
For doxygen, there is a requirement from nntrainer repository that
doxygen function definition should not be checked if the function
definition is inside a source.

This patch implments for the requirement as an option.
`pr_doxygen_check_skip_function_definition` if this flag is turned on,
doxygen checker does not check definitions inside a sourcefile(eg,
*.cpp)

Feature:
- Add option to skip definition for non header file

Additional Fix:
- Fix a doxygen checker to check the function declaration(prototype)

**How test was done**
- Apply to nntrainer repository.
- Check doxygen book is not uglyfied due to this change.

**Self evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test: [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: Jihoon Lee <jhoon.it.lee@samsung.com>